refactor(test_auxiliary.py): separating the contents of the auxiliary tests
agora os testes estão separados entre unitarios e de integração
